+++ src/main/java/com/cells/network/CellsNetworkHandler.java
@@ -6,6 +6,7 @@ import com.cells.network.packets.PacketQuickAddFluidFilter;
 import com.cells.network.packets.PacketQuickAddItemFilter;
 import com.cells.network.packets.PacketSaveMemoryCardWithFilters;
 import com.cells.network.packets.PacketSetConfigurableCellCapacity;
+import com.cells.network.packets.PacketSetConfigurableCellMaxTypes;
 import com.cells.network.packets.PacketSetMaxSlotSize;
 import com.cells.network.packets.PacketSetPollingRate;
 
@@ -28,6 +29,7 @@ public class CellsNetworkHandler {
         INSTANCE.registerMessage(PacketOpenGui.Handler.class, PacketOpenGui.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketSetPollingRate.Handler.class, PacketSetPollingRate.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketSetConfigurableCellCapacity.Handler.class, PacketSetConfigurableCellCapacity.class, packetId++, Side.SERVER);
+        INSTANCE.registerMessage(PacketSetConfigurableCellMaxTypes.Handler.class, PacketSetConfigurableCellMaxTypes.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketQuickAddItemFilter.Handler.class, PacketQuickAddItemFilter.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketQuickAddFluidFilter.Handler.class, PacketQuickAddFluidFilter.class, packetId++, Side.SERVER);
         INSTANCE.registerMessage(PacketSaveMemoryCardWithFilters.Handler.class, PacketSaveMemoryCardWithFilters.class, packetId++, Side.SERVER);